Create Auxiliary Fantasy Stats Tables for weekly and yearly summaries
create_weekly_stats first creates a yearly_stats table for the main
purpose of identifying every player that recorded a stat for a
particular season. The yearly_stats table is then iterated and linked
to a series of records keyed by player_id / week_number in a
weekly_stats table. This table is representation of each week played
during the season by a player in the yearly_stats table. The
create_weekly_stretch table simply creates a weekly_stretch_table that
is keyed by player_id, week1 (the start week of a 4 game stretch).  The
total fantasy points for a 4 week stretch, and later any N week stretch
is computed per record. If a player misses a week due to injury, bye
week, or any other reason, that particular week is excluded from the
stretch.  The purpose of this table is to identify the best N week
stretches.